pyproject.toml
[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[project]
name = "teammindhub-backend"
version = "0.1.0"
description = "Local-first FastAPI backend for TeamMindHub."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "httpx>=0.27.0,<1.0.0",
  "python-multipart>=0.0.9,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "pywebview>=5.3.2,<6.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
]
desktop-build = [
  "pyinstaller>=6.11.0,<7.0.0",
]
oss = [
  "docling>=2.0.0,<3.0.0",
  "langgraph>=0.2.0,<1.0.0",
  "llama-index-core>=0.14.0,<1.0.0",
  "llama-index-retrievers-bm25>=0.3.0,<1.0.0",
  "ragflow-sdk>=0.24.0,<1.0.0",
]

[project.scripts]
teammindhub-desktop = "app.desktop.main:main"
teammindhub-desktop-build = "app.desktop.build:main"
teammindhub-desktop-doctor = "app.desktop.doctor:main"

[tool.pytest.ini_options]
addopts = "-p no:cacheprovider"
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]
